requirements.txt
Pillow==7.1.1
pyzbar==0.1.8
pypng==0.0.20
qrcode==6.1
